Refactor InferenceEngine and PagedKv components for improved performance and clarity
- Removed unused configuration parameters related to page-store compaction in InferenceEngine.
- Enhanced error handling during KV swap-in operations, ensuring failures are reported.
- Introduced device token input handling for batch decoding, allowing for more efficient token management.
- Updated statistics tracking to include metrics for device token input hits and tokens processed.
- Simplified the logic for gathering and materializing paged KV data, improving clarity and maintainability.
- Added new methods for managing paged KV rows and handling batch setups, ensuring better performance in mixed batch scenarios.
- Adjusted sampling logic to accommodate device tokens, enhancing the sampling process.
- Cleaned up unused code and added necessary documentation for clarity.

+++ crane-core/kernels/fused_ops/sampling.cuh
@@ -148,7 +148,7 @@ extern "C" __global__ void gpu_argmax_phase2(
 
 extern "C" __global__ void gpu_argmax_batch_bf16(
     const __nv_bfloat16 *__restrict__ logits,  // [batch_size, vocab_size]
-    int32_t             *__restrict__ output_tokens,
+    uint32_t            *__restrict__ output_tokens,
     const int batch_size,
     const int vocab_size
 ) {
@@ -206,7 +206,7 @@ extern "C" __global__ void gpu_argmax_batch_bf16(
         }
 
         if (lane_id == 0) {
-            output_tokens[row] = local_idx;
+            output_tokens[row] = (uint32_t)local_idx;
         }
     }
 }
